refactor: extract portfolio shrinkage into experiment_utils.shrinkage
Move winner's-curse / EB / cumulative / joint helpers out of utils, add
NSS-adjusted cumulative impact, and update docs and examples.
