+++ examples/cumulative_methods_recovery.py
@@ -14,20 +14,18 @@
 # 7. One-sided `winners_curse_estimate` is median-unbiased among winners
 
 # %%
-from __future__ import annotations
-
 import numpy as np
 from scipy.stats import norm
 
-from experiment_utils import (
+from experiment_utils.shrinkage import (
     cumulative_impact,
+    empirical_bayes_shrinkage,
     estimate_guardrail_rho,
     fit_normal_prior_map,
     joint_metric_shrinkage,
     process_level_total_effect,
     winners_curse_estimate,
 )
-from experiment_utils.utils import empirical_bayes_shrinkage
 
 PASS = "PASS"
 FAIL = "FAIL"
